Widen CSRMM index arithmetic to size_t
The CSRMM kernels addressed the dense operands with 32-bit arithmetic:

    B[indices[j] * n + c]
    C[row * n + c]

`indices[j]` is int32 and `n` is int, so the product is computed in 32 bits and
wraps once the dense operand exceeds INT32_MAX elements. Most of the package
already widens these (binary_fcnmm.cu, binary_densemv.cu, the jit families all
cast to size_t); the three CSRMM files did not.

This is reachable, not theoretical: `B` is usually a bool/int8 event matrix, so
2 GiB of allocation is enough to cross the boundary. Demonstrated on the GPU
with k=2098152, n=1024 (k*n = 2148507648):

    indices[0] = 2098151, n = 1024, c = 0
      int    arithmetic -> -2146460672   <-- negative, out-of-bounds read
      size_t arithmetic ->  2148506624   correct

Widened 30 subscripts across binary_csrmm.cu, binary_indexed_csrmm.cu and
float_csrmm.cu. The `smem[strip * 32 + lane]` subscripts are deliberately left
alone: they are bounded by the block size (256 slots) and cannot overflow.

All three files compile-checked with nvcc; the full GPU CSR suite passes
(699 passed).

+++ brainevent/_csr/binary_csrmm.cu
@@ -79,16 +79,16 @@ __global__ void _csrmm_nt_warp_homo_kern##SUFFIX(                              \
         IndptrT j = start;                                                         \
         _Pragma("unroll 4")                                                    \
         for (; j + 1 < end; j += 2) {                                          \
-            ACC_T mask0 = (ACC_T)IS_ACTIVE(B[indices[j]   * n + c]);           \
-            ACC_T mask1 = (ACC_T)IS_ACTIVE(B[indices[j+1] * n + c]);          \
+            ACC_T mask0 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j]   * n + c]);   \
+            ACC_T mask1 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j+1] * n + c]);  \
             acc0 += w * mask0;                                                 \
             acc1 += w * mask1;                                                 \
         }                                                                      \
         for (; j < end; j++) {                                                 \
-            ACC_T mask = (ACC_T)IS_ACTIVE(B[indices[j] * n + c]);              \
+            ACC_T mask = (ACC_T)IS_ACTIVE(B[(size_t)indices[j] * n + c]);      \
             acc0 += w * mask;                                                  \
         }                                                                      \
-        C[row * n + c] = WRITE_W(acc0 + acc1);                                 \
+        C[(size_t)row * n + c] = WRITE_W(acc0 + acc1);                         \
     }                                                                          \
 }
 
@@ -117,13 +117,13 @@ __global__ void _csrmm_nt_block_homo_kern##SUFFIX(                              
             IndptrT j = start + strip;                                              \
             _Pragma("unroll 4")                                                \
             for (; j + 8 < end; j += 16) {                                     \
-                ACC_T mask0 = (ACC_T)IS_ACTIVE(B[indices[j]   * n + c]);       \
-                ACC_T mask1 = (ACC_T)IS_ACTIVE(B[indices[j+8] * n + c]);      \
+                ACC_T mask0 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j]   * n + c]); \
+                ACC_T mask1 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j+8] * n + c]); \
                 acc0 += w * mask0;                                              \
                 acc1 += w * mask1;                                              \
             }                                                                  \
             for (; j < end; j += 8) {                                           \
-                ACC_T mask = (ACC_T)IS_ACTIVE(B[indices[j] * n + c]);          \
+                ACC_T mask = (ACC_T)IS_ACTIVE(B[(size_t)indices[j] * n + c]);  \
                 acc0 += w * mask;                                               \
             }                                                                  \
         }                                                                       \
@@ -133,7 +133,7 @@ __global__ void _csrmm_nt_block_homo_kern##SUFFIX(                              
             ACC_T sum = ACC_ZERO;                                               \
             _Pragma("unroll 8")                                                \
             for (int s = 0; s < 8; s++) sum += smem[s * 32 + lane];            \
-            C[row * n + c] = WRITE_W(sum);                                     \
+            C[(size_t)row * n + c] = WRITE_W(sum);                             \
         }                                                                       \
         __syncthreads();                                                        \
     }                                                                           \
@@ -169,16 +169,16 @@ __global__ void _csrmm_nt_warp_hetero_kern##SUFFIX(                             
         IndptrT j = start;                                                           \
         _Pragma("unroll 4")                                                      \
         for (; j + 1 < end; j += 2) {                                            \
-            ACC_T mask0 = (ACC_T)IS_ACTIVE(B[indices[j]   * n + c]);             \
-            ACC_T mask1 = (ACC_T)IS_ACTIVE(B[indices[j+1] * n + c]);            \
+            ACC_T mask0 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j]   * n + c]);     \
+            ACC_T mask1 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j+1] * n + c]);    \
             acc0 += READ_W(weights[j])   * mask0;                                \
             acc1 += READ_W(weights[j+1]) * mask1;                                \
         }                                                                        \
         for (; j < end; j++) {                                                   \
-            ACC_T mask = (ACC_T)IS_ACTIVE(B[indices[j] * n + c]);                \
+            ACC_T mask = (ACC_T)IS_ACTIVE(B[(size_t)indices[j] * n + c]);        \
             acc0 += READ_W(weights[j]) * mask;                                   \
         }                                                                        \
-        C[row * n + c] = WRITE_W(acc0 + acc1);                                   \
+        C[(size_t)row * n + c] = WRITE_W(acc0 + acc1);                           \
     }                                                                            \
 }
 
@@ -206,13 +206,13 @@ __global__ void _csrmm_nt_block_hetero_kern##SUFFIX(                            
             IndptrT j = start + strip;                                                \
             _Pragma("unroll 4")                                                  \
             for (; j + 8 < end; j += 16) {                                       \
-                ACC_T mask0 = (ACC_T)IS_ACTIVE(B[indices[j]   * n + c]);         \
-                ACC_T mask1 = (ACC_T)IS_ACTIVE(B[indices[j+8] * n + c]);        \
+                ACC_T mask0 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j]   * n + c]); \
+                ACC_T mask1 = (ACC_T)IS_ACTIVE(B[(size_t)indices[j+8] * n + c]); \
                 acc0 += READ_W(weights[j])   * mask0;                            \
                 acc1 += READ_W(weights[j+8]) * mask1;                            \
             }                                                                    \
             for (; j < end; j += 8) {                                             \
-                ACC_T mask = (ACC_T)IS_ACTIVE(B[indices[j] * n + c]);            \
+                ACC_T mask = (ACC_T)IS_ACTIVE(B[(size_t)indices[j] * n + c]);    \
                 acc0 += READ_W(weights[j]) * mask;                               \
             }                                                                    \
         }                                                                         \
@@ -222,7 +222,7 @@ __global__ void _csrmm_nt_block_hetero_kern##SUFFIX(                            
             ACC_T sum = ACC_ZERO;                                                 \
             _Pragma("unroll 8")                                                  \
             for (int s = 0; s < 8; s++) sum += smem[s * 32 + lane];              \
-            C[row * n + c] = WRITE_W(sum);                                       \
+            C[(size_t)row * n + c] = WRITE_W(sum);                               \
         }                                                                         \
         __syncthreads();                                                          \
     }                                                                             \
